services/resource_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pytest
pytest-cov
flake8
bandit
redis
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
python-dotenv
PyJWT 